+++ packages/donate-button-v4/src/components/widget/components/PaymentProcess/CryptoFlow/index.tsx
@@ -1,0 +1,25 @@
+import {
+	LargePaymentMethodSelect,
+	SmallPaymentMethodSelect
+} from 'src/components/widget/components/PaymentProcess/PaymentMethodSelect';
+import {RedirectNotice} from 'src/components/widget/components/PaymentProcess/RedirectNotice';
+import {
+	formCss,
+	formContainerCss
+} from 'src/components/widget/components/PaymentProcess/styles';
+import {useSubmitDonation} from 'src/components/widget/hooks/useSubmitDonation';
+
+export const CryptoFlow = () => {
+	const submitDonation = useSubmitDonation();
+
+	return (
+		<form className={formCss} onSubmit={submitDonation}>
+			<LargePaymentMethodSelect />
+			<div className={formContainerCss}>
+				<SmallPaymentMethodSelect />
+				<div>CryptoFlow</div>
+				<RedirectNotice />
+			</div>
+		</form>
+	);
+};
