+++ packages/donate-button-v4/src/components/widget/components/PaymentProcess/styles.ts
@@ -1,7 +1,7 @@
 import cxs from 'cxs';
 import {BREAKPOINTS} from 'src/components/widget/theme/breakpoints';
 import {textSize} from 'src/components/widget/theme/font-sizes';
-import {Spacing} from 'src/components/widget/theme/spacing';
+import {Spacing, verticalStackCss} from 'src/components/widget/theme/spacing';
 
 export const cardCss = cxs({
 	padding: `${Spacing.XL} ${Spacing.XL} 0px`,
@@ -9,6 +9,7 @@ export const cardCss = cxs({
 		padding: '0px'
 	}
 });
+
 export const fieldSetCss = cxs({
 	border: 'none',
 	padding: 0,
@@ -22,6 +23,14 @@ export const legendCss = cxs({
 	fontWeight: 700,
 	...textSize.s,
 	marginBottom: Spacing.S
+});
+
+export const formContainerCss = cxs({
+	...verticalStackCss.cxs(Spacing.XL),
+	padding: '0px',
+	[BREAKPOINTS.TabletLandscapeUp]: {
+		padding: Spacing.XXL
+	}
 });
 
 export const formCss = cxs({
